requirements.txt
#
# Copyright (C) 2021 Vaticle
#
# Licensed to the Apache Software Foundation (ASF) under one
# or more contributor license agreements.  See the NOTICE file
# distributed with this work for additional information
# regarding copyright ownership.  The ASF licenses this file
# to you under the Apache License, Version 2.0 (the
# "License"); you may not use this file except in compliance
# with the License.  You may obtain a copy of the License at
#
#   http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing,
# software distributed under the License is distributed on an
# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
# KIND, either express or implied.  See the License for the
# specific language governing permissions and limitations
# under the License.
#

## To install all dependencies, run: pip3 install -r requirements.txt


## Configuration options


# Allow importing of snapshots
--extra-index-url https://repo.vaticle.com/repository/pypi-snapshot/simple


## Dependencies

# IMPORTANT: Any changes to these dependencies should be copied to requirements_dev.txt.
typedb-protocol==0.0.0-d608bb39f21013ed0e569581b599900d1d60ede2
grpcio==1.36.1
protobuf==3.15.5
